feat: "방 전체 목록 조회 API 구현"

+++ src/main/java/my/tableTennisGame/web/dto/user/UserRespDto.java
@@ -20,7 +20,7 @@ public class UserRespDto {
 
         public UserListRespDto(Page<User> userPage) {
             this.totalElements = (int) userPage.getTotalElements();
-            this.totalPages = (int) userPage.getTotalPages();
+            this.totalPages = userPage.getTotalPages();
             this.userList = userPage.getContent().stream()
                     .map(UserDto::new).toList();
         }
